Refactor SentinelOne ingestion to use common job parameters
### Summary
Updated the SentinelOne ingestion process to utilize a common job parameters dictionary, improving code organization and readability. The API URL and token are now passed through this dictionary, and a new `api.py` file has been created to encapsulate API call logic, replacing the previous utility functions in `utils.py`. This change enhances maintainability and prepares the codebase for future expansions.

### Changes
- Introduced `common_job_parameters` in `__init__.py` for better parameter management.
- Refactored `sync_accounts` to use parameters from the common dictionary.
- Moved API call logic from `utils.py` to a new `api.py` file for better separation of concerns.
- Updated integration tests to reflect changes in parameter handling.
