+++ tests/integration/cartography/intel/sentinelone/test_account.py
@@ -22,14 +22,13 @@ def test_sync_account(neo4j_session, mocker):
     # Create common job parameters required by the new sync function
     common_job_parameters = {
         "UPDATE_TAG": TEST_UPDATE_TAG,
+        "API_URL": "https://test-api.sentinelone.net",
+        "API_TOKEN": "test-api-token",
     }
 
     account_ids = [ACCOUNT_ID, ACCOUNT_ID_2]
     synced_account_ids = sync_accounts(
         neo4j_session,
-        "https://test-api.sentinelone.net",
-        "test-api-token",
-        TEST_UPDATE_TAG,
         common_job_parameters,
         account_ids,
     )
